refactor: HomeContentItem을 SoptContentItem으로 공통 컴포넌트로 변경
